move RequireAuth to its own file
