+++ src/components/MainLayout/MainLayout.tsx
@@ -1,22 +1,14 @@
-import { ReactElement, Suspense, useEffect, useState } from 'react';
-
-import { useWeb3React } from '@web3-react/core';
+import { Suspense, useState } from 'react';
 
 import HelpButton from '../HelpButton';
-import {
-  GlobalUi,
-  LoadingModal,
-  SentryScopeController,
-  WalletAuthModal,
-} from 'components';
-import { useApiBase } from 'hooks';
-import { AuthContext, useAuthStep } from 'hooks/login';
+import { GlobalUi, SentryScopeController } from 'components';
 import type { AuthStep } from 'hooks/login';
-import { useWalletAuth } from 'recoilState';
+import { AuthContext } from 'hooks/login';
 import { AppRoutes } from 'routes/routes';
 import { Box } from 'ui';
 
 import { MainHeader } from './MainHeader';
+import { RequireAuth } from './RequireAuth';
 
 // this component sets up the top navigation bar to stay fixed on-screen and
 // have content scroll underneath it
@@ -67,39 +59,4 @@ export default MainLayout;
 // this is in this file because it depends on the <main> tag defined above
 export const scrollToTop = () => {
   document.getElementsByTagName('main')[0].scrollTop = 0;
-};
-
-const RequireAuth = (props: { children: ReactElement }) => {
-  const address = useWalletAuth().address;
-  const web3Context = useWeb3React();
-  const { finishAuth } = useApiBase();
-  const [authStep, setAuthStep] = useAuthStep();
-
-  useEffect(() => {
-    // reset after logging out or signature error
-    if (authStep !== 'connect' && !web3Context.active) {
-      setAuthStep('connect');
-      return;
-    }
-
-    if (authStep === 'connect' && web3Context.active) {
-      setAuthStep('sign');
-      finishAuth({ web3Context }).then(success => {
-        if (!success) {
-          web3Context.deactivate();
-        } else {
-          setAuthStep('done');
-        }
-      });
-    }
-  }, [address, web3Context]);
-
-  // step 1: get a wallet connection
-  if (authStep === 'connect') return <WalletAuthModal open />;
-
-  // step 2: reuse an auth token, or get a new one with a signature
-  if (authStep !== 'done') return <LoadingModal visible note="RequireAuth" />;
-
-  // step 3: render routes
-  return props.children;
 };
